Ajout de verif si 0 < note < 20 et modificiation du type de la note : int à double

+++ Diplodocus/Models/SchoolSubjectMark.cs
@@ -11,7 +11,7 @@
 
         [Key]
         public int IdMark { get; set; }
-        public Nullable<int> Mark { get; set; }
+        public Nullable<double> Mark { get; set; }
         public virtual Student Student { get; set; }
         public virtual SchoolSubject SchoolSubject { get; set; }
         public int SchoolSubjectIdSchoolSubject { get; set; }
